requirements.txt
matplotlib==3.2.2
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_learn==1.1.2
torch==1.10.0
torchvision==0.11.1
